% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_amino_acid_percentages.R
\name{amino_acid_percentages}
\alias{amino_acid_percentages}
\title{Calculate aminoacid frequencies}
\usage{
amino_acid_percentages(input)
}
\arguments{
\item{input}{either a vector of peptides or a data frame with a variable
called "window_seq" containing the peptides.}
}
\value{
Data frame containing the calculated AA percentages. If `input` is
a`data.frame` the original input is returned with the AA percentages
appended as columns.
}
\description{
This function is used to calculate aminoacid frequencies from a vector of
peptides.
}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk});
        Jodie Ashford (\email{ashfojsm@aston.ac.uk})
}
