% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_calc_aa_composition.R
\name{calc_aa_composition}
\alias{calc_aa_composition}
\title{Calculate qualitative aminoacid composition of peptides}
\usage{
calc_aa_composition(df, cl)
}
\arguments{
\item{df}{a \emph{data.table} of class \emph{windowed_epit_dt} or \emph{windowed_prot_dt},
generated by \code{\link[=make_window_df]{make_window_df()}}}

\item{cl}{either a cluster object or a positive integer.}
}
\value{
updated \strong{df} data.table containing the AA percentage features
appended as columns.
}
\description{
This function is used to calculate the percent composition of peptides, in
terms of nine AA types: "Tiny", "Small", "Aliphatic", "Aromatic",
"NonPolar", "Polar", "Charged", "Basic" and "Acidic".
}
\details{
This function is a direct adaptation of \code{\link[Peptides:aacomp]{Peptides::aaComp()}}
}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk});
Jodie Ashford (\email{ashfojsm@aston.ac.uk})
}
