% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cojoint_triads.R
\name{calc_cojoint_triads}
\alias{calc_cojoint_triads}
\title{Calculate cojoint triads}
\usage{
calc_cojoint_triads(input, ncores = parallel::detectCores() - 1)
}
\arguments{
\item{input}{either a vector of peptides or a data frame with a variable
called "window_seq" containing the peptides.}

\item{ncores}{number of cores to use when calculating the features.}
}
\value{
Data frame containing the calculated AA percentages. If `input` is
a`data.frame` the original input is returned with the AA percentages
appended as columns.
}
\description{
This function is used to calculate the cojoint triads for a vector of
peptides.
}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk});
        Jodie Ashford (\email{ashfojsm@aston.ac.uk})
}
