% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_calc_conjoint_triads.R
\name{calc_conjoint_triads}
\alias{calc_conjoint_triads}
\title{Calculate conjoint triads}
\usage{
calc_conjoint_triads(df, cl)
}
\arguments{
\item{df}{a \emph{data.table} of class \emph{windowed_epit_dt} or \emph{windowed_prot_dt},
generated by \code{\link[=make_window_df]{make_window_df()}}}

\item{cl}{either a cluster object or a positive integer.}
}
\value{
updated \strong{df} data.table containing the conjoint triad frequencies
}
\description{
This function is used to calculate the conjoint triad frequencies for
each peptide
}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk});
Jodie Ashford (\email{ashfojsm@aston.ac.uk})
}
