% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_molecular_weight.R
\name{calc_molecular_weight}
\alias{calc_molecular_weight}
\title{Calculate molecular weight of peptide sequences}
\usage{
calc_molecular_weight(input)
}
\arguments{
\item{input}{either a vector of peptides or a data frame with a variable
called "window_seq" containing the peptides.}
}
\value{
Data frame containing the calculated molecular weights. If \code{input} is
a\code{data.frame} the original input is returned with the molecular weights
appended as columns.
}
\description{
This function is used to calculate the molecular weights from a vector
of peptides.
}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk});
Jodie Ashford (\email{ashfojsm@aston.ac.uk})
}
