% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_Npeptide_composition.R
\name{calc_Npeptide_composition}
\alias{calc_Npeptide_composition}
\title{Calculate frequencies of N-peptides}
\usage{
calc_Npeptide_composition(df, N, cl)
}
\arguments{
\item{df}{a \emph{data.table} of class \emph{windowed_epit_dt} or \emph{windowed_prot_dt},
generated by \code{\link[=make_window_df]{make_window_df()}}}

\item{N}{the length of the subsequences to consider.}

\item{cl}{either a cluster object or a positive integer.}
}
\value{
updated \strong{df} data.table containing the N-peptide frequencies
}
\description{
This function is used to calculate the frequencies of N-peptide subsequences
for each peptide
}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk});
Jodie Ashford (\email{ashfojsm@aston.ac.uk})
}
