% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_number_of_atoms.R
\name{calc_number_of_atoms}
\alias{calc_number_of_atoms}
\title{Calculate number of each type of atom in peptide sequences}
\usage{
calc_number_of_atoms(df, cl)
}
\arguments{
\item{df}{a \emph{data.table} of class \emph{windowed_epit_dt} or \emph{windowed_prot_dt},
generated by \code{\link[=make_window_df]{make_window_df()}}}

\item{cl}{either a cluster object or a positive integer.}
}
\value{
updated \strong{df} data.table containing the number of atoms of elements
C, H, O, N and S.
}
\description{
This function is used to calculate the number of specific atoms in each
peptide
}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk});
Jodie Ashford (\email{ashfojsm@aston.ac.uk})
}
