% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_peptide_features.R
\name{calc_peptide_features}
\alias{calc_peptide_features}
\title{Calculate simple features based on composition and physico-chemical
properties of a protein sequence}
\usage{
calc_peptide_features(seqs, .funs = c(mean, sd), ncpus = 1,
  save_folder = NULL)
}
\arguments{
\item{seqs}{A character vector of potein sequences.}

\item{.funs}{a vector of functions used to aggregate some of the AA-based
features into sequence-wide features (see `Value`).}

\item{ncpus}{positive integer, number of cores to use (multi-core
capabilities not yet available for Windows systems.)}

\item{save_folder}{path to folder for saving the results.}
}
\value{
A data frame containing the extracted features is returned invisibly.
This function currently calculates the following features for each sequence
provided in `seqs`:

  \itemize{
     \item\code{AA composition}, based on function `Peptides::aaComp()`.
     Calculates the percent of amino acids by class (Tiny, Small, Aliphatic,
     Aromatic, Non-polar, Polar, Charged, Basic and Acidic) in the sequence.
     \item\code{AA descriptors}, based on function
     `Peptides::aaDescriptors()`. For each sequence the following properties
     are calculated for each AA, and then aggregated using the functions
     passed in `.funs`.

  \itemize{
     \item\code{crucianiProperties}
     \itemize{
         \item PP1: Polarity
         \item PP2: Hydrophobicity
         \item PP3: H-bonding
     }
     \item\code{kideraFactors} (the four first features are essentially
     pure physical properties; the remaining six are linear combinations of
     several properties):
     \itemize{
         \item KF1: Helix/bend preference
         \item KF2: Side-chain size
         \item KF3: Extended structure preference
         \item KF4: Hydrophobicity
         \item KF5 - KF10: linear combinations of other characteristics
     }
     \item\code{zScales} (based on physicochemical properties of the AAs
     including NMR data and thin-layer chromatography data):
     \itemize{
         \item Z1: Lipophilicity
         \item Z2: Steric properties (Steric bulk/Polarizability)
         \item Z3: Electronic properties (Polarity / Charge)
         \item Z4-Z5: relate electronegativity, heat of formation,
         electrophilicity and hardness.
     }
     \item\code{FASGAI} (based on physicochemical properties of the AAs
     including NMR data and thin-layer chromatography data):
     \itemize{
         \item F1: Hydrophobicity index
         \item F2: Alpha and turn propensities
         \item F3: Bulky properties
         \item F4: Compositional characteristic index
         \item F5: Local flexibility
         \item F6: Electronic properties
     }
     \item\code{tScales} (based on 67 common topological descriptors of
     amino acids. These topological descriptors are based on the connectivity
     table of amino acids alone, and to not explicitly consider 3D properties
     of each structure):
     \itemize{
         \item T1 - T5
     }
     \item\code{VHSE scales} (principal component score Vectors of
     Hydrophobic, Steric, and Electronic properties. Derived from PCA on
     independent families of 18 hydrophobic properties, 17 steric properties,
     and 15 electronic properties):
     \itemize{
         \item VHSE1 - VHSE2: Hydrophobic properties
         \item VHSE3 - VHSE4: Steric properties
         \item VHSE5 - VHSE8: Electronic properties
     }
     \item\code{ProtFP descriptors}: (constructed from a large initial
     selection of indices obtained from the AAindex database for all 20
     naturally occurring amino acids.):
     \itemize{
         \item ProtFP1-ProtFP8
     }
     \item\code{stScales} (proposed by Yang et al.'2010, taking 827
     properties into account which are mainly constitutional, topological,
     geometrical, hydrophobic, electronic, and steric properties of AAs):
     \itemize{
         \item ST1 - ST8
     }
     \item\code{BLOSUM indices} (derived of physicochemical properties that
     have been subjected to a VARIMAX analyses and an alignment matrix of the
     20 natural AAs using the BLOSUM62 matrix):
     \itemize{
         \item BLOSUM1 - BLOSUM10
     }
     \item\code{MS-WHIM scores} (derived from 36 electrostatic potential
     properties derived from the three-dimensional structure of the 20
     natural amino acids):
     \itemize{
         \item MSWHIM1 - MSWHIM3
     }
  }
}

See the documentation for `Peptides::aaDescriptors()` and
`Peptides::aaData()` for further information and references.
}
\description{
This function relies on package `Peptides` to calculate several simple
features. It currently
}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk})
}
