% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_performance_perAA.R
\name{calc_performance_perAA}
\alias{calc_performance_perAA}
\title{Calculate epitope classification performance indices by individual position}
\usage{
calc_performance_perAA(truth, pred, posValue = 1, negValue = -1)
}
\arguments{
\item{truth}{vector of reference values}

\item{pred}{vector of predicted values}

\item{posValue}{value that indicates the "positive" class.}

\item{negValue}{value that indicates the "negative" class.}
}
\value{
list containing several performance indicators
}
\description{
This function is used to calculate several distinct BINARY classification
performance indicators on a position-by-position manner (in contrast to a
'by-labelled-region' approach). \code{NA} values in either input vector are not
considered in the calculations.
}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk});
Jodie Ashford (\email{ashfojsm@aston.ac.uk})
}
