% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sequence_entropy.R
\name{calc_sequence_entropy}
\alias{calc_sequence_entropy}
\title{Calculate sequence entropy of peptides}
\usage{
calc_sequence_entropy(input)
}
\arguments{
\item{input}{either a vector of peptides or a data frame with a variable
called "window_seq" containing the peptides.}
}
\value{
Data frame containing the calculated AA percentages. If \code{input} is
a\code{data.frame} the original input is returned with the AA percentages
appended as columns.
}
\description{
This function is used to calculate the sequence entropies
from a vector of peptides. Entropy = -sum(p_k * log2(p_k))
}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk});
Jodie Ashford (\email{ashfojsm@aston.ac.uk})
}
