% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate_data.R
\name{consolidate_data}
\alias{consolidate_data}
\title{Consolidate epitope and protein data}
\usage{
consolidate_data(
  epitopes,
  proteins,
  save_folder = NULL,
  only_exact = FALSE,
  set.positive = c("any", "mode", "all"),
  ncpus = 1
)
}
\arguments{
\item{epitopes}{data frame of epitope data (returned by \code{\link[=get_LBCE]{get_LBCE()}}).}

\item{proteins}{data frame of protein data (returned by \code{\link[=get_proteins]{get_proteins()}}).}

\item{save_folder}{path to folder for saving the results.}

\item{only_exact}{logical, should only sequences labelled as "Exact Epitope"
in variable \emph{epit_struc_def} (within \code{epitopes}) be considered?}

\item{set.positive}{how to decide whether an observation should be of the
"Positive" (+1) or "Negative" (-1) class? Use "any" to set a sequence as positive if
$n_positive > 0$, "mode" to set it if $n_positive >= n_negative$,
or "all" to set it if $n_negative == 0$. Defaults to "mode".}

\item{ncpus}{positive integer, number of cores to use}
}
\value{
A data frame containing all proteins mentioned that appear at least
once in \code{epitopes$protein_id} (one row per aminoacid residue), containing
the consolidated information extracted from \code{epitopes}.
}
\description{
Label protein positions based on linear B-cell epitope information extracted
from IEDB.
}
\details{
This routine is used to consolidate epitope information, by:
\itemize{
\item Filtering \code{proteins} such that only entries that appear at least
once in \code{epitopes$protein_id} are retained
\item Using the \code{epitopes} data (fields \code{protein_id}, \code{start_pos} and
\code{end_pos}) to generate labels for each position of the selected proteins.
}

Prior to the consolidation, entries in \code{epitopes} input are removed if they:
\itemize{
\item Lack a valid \emph{protein_id} (i.e., if they have no corresponding
entry in \code{proteins})
\item Lack a valid string in \emph{epit_seq}
\item Lack a valid definition in \emph{epit_struc_def}
\item Have a mismatch between the sequence in \emph{epit_seq} and the
corresponding sequence between \emph{start_pos} and \emph{end_pos} on the protein
sequence.
}

Entries in the \code{proteins} data frame with a value of \emph{TSeq_seqtype} different
from "protein" are also removed.
}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk})
}
