% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_get_IEDB.R
\name{get_IEDB}
\alias{get_IEDB}
\title{Download and unzip IEDB database (XML export)}
\usage{
get_IEDB(
  url = "https://www.iedb.org/downloader.php?file_name=doc/iedb_export.zip",
  save_folder = NULL,
  remove_zip = TRUE
)
}
\arguments{
\item{url}{URL of the \emph{.zip} file containing the full IEDB
export (XML).}

\item{save_folder}{Path to folder for extracting the results. Defaults to
"iedb_yyyymmdd", where \emph{yyyymmdd} is replaced by the current date.}

\item{remove_zip}{logical flag: should the \emph{.zip} file be deleted after
extraction?}
}
\description{
This function is used to retrieve the full IEDB database from
\href{https://www.iedb.org}{IEDB} and extract it to a target folder.
}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk})
}
