% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_deprecated.R
\name{get_linear_bcell_epitopes}
\alias{get_linear_bcell_epitopes}
\title{Extract linear B Cell epitopes from XML files retrieved from IEDB.}
\usage{
get_linear_bcell_epitopes(data_folder = "./", save_folder = NULL)
}
\arguments{
\item{data_folder}{path (either relative or absolute) to the directory
containing the XML files}

\item{save_folder}{path to folder for saving the results.}
}
\description{
This function is deprecated. Please use get_LBCE() instead.
}
