% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_proteins.R
\name{get_proteins}
\alias{get_proteins}
\title{Retrieve protein sequences and data from GenBank}
\usage{
get_proteins(uids, ncpus = 1, save_folder = NULL)
}
\arguments{
\item{uids}{A list of potein IDs provided a character vector.}

\item{ncpus}{positive integer, number of cores to use (multi-core
capabilities not yet available for Windows systems.)}

\item{save_folder}{path to folder for saving the results.}
}
\value{
A data frame containing the extracted proteins is returned invisibly.
}
\description{
This function is used to retrieve data from
[GenBank](https://www.ncbi.nlm.nih.gov/genbank/) for given protein IDs.
}
\author{
Felipe Campelo (\email{fcampelo@ufmg.br},
\email{f.campelo@aston.ac.uk})
}
