% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_get_taxonomy.R
\name{get_taxonomy}
\alias{get_taxonomy}
\title{Retrieve taxonomic classification tables from NCBI.}
\usage{
get_taxonomy(uids, save_folder = NULL)
}
\arguments{
\item{uids}{vector of organism IDs to retrieve taxonomical information, or a
\emph{data.table} object of class \emph{LBCE_dt}, generated by \code{\link[=get_LBCE]{get_LBCE()}}.}

\item{save_folder}{path to folder for saving the results.}
}
\value{
A list containing the information for each element of uids
}
\description{
This function retrieves the taxonomic information of a vector of organism
IDs, from the NCBI Taxonomy data base.
}
\examples{
uids <- c("6282", # O. volvulus,
          "9606") # H. sapiens
tax <- get_taxonomy(uids)

}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk})
}
