% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_proteins.R
\name{label_proteins}
\alias{label_proteins}
\title{Label protein positions based on epitope data}
\usage{
label_proteins(
  proteins,
  epitopes,
  set.positive = c("any", "mode", "all"),
  ncpus = 1,
  save_folder = NULL
)
}
\arguments{
\item{proteins}{data frame of protein data, returned by \code{\link[=get_proteins]{get_proteins()}} or
\code{\link[=make_window_df]{make_window_df()}}.}

\item{epitopes}{data frame of epitope data, returned by \code{\link[=prepare_join_df]{prepare_join_df()}}
or \code{\link[=filter_epitopes]{filter_epitopes()}}.}

\item{set.positive}{how to decide whether a position should be labeled as
"Positive" (+1). Use "any" to set a position as positive if
it is labeled as $+1$ in at least one entry of \strong{epitopes}; "mode" to
set it by majority voting; or "all" to only label a position as
Positive if it has at least one occurrence as $+1$ and none as $-1$.
Unlabelled positions receive \strong{NA} in their \emph{Class} column.}

\item{ncpus}{number of cores to use}

\item{save_folder}{path to folder for saving the results.}
}
\value{
A data table of class \emph{windowed_prot_dt} with columns containing the
number of positive and negative labels found for each position of each
protein, plus a \emph{Class} column calculated according to \emph{set.positive}.
}
\description{
Takes a data.table of data of class \emph{protein_dt} (returned by
\code{\link[=get_proteins]{get_proteins()}}) or \emph{windowed_prot_dt} (returned by \code{\link[=make_window_df]{make_window_df()}}) and
labels each protein position based on known epitope/non-epitope regions
documented in an \strong{epitopes} data.table of class \emph{joined_epit_dt} (returned
by \code{\link[=prepare_join_df]{prepare_join_df()}} or by \code{\link[=filter_epitopes]{filter_epitopes()}}).
}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk})
}
