% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_proteins.R
\name{label_proteins2}
\alias{label_proteins2}
\title{Label protein positions based on epitope data}
\usage{
label_proteins2(proteins, epitopes, ncpus = 1)
}
\arguments{
\item{proteins}{data frame of protein data.}

\item{epitopes}{data frame of epitope data.}

\item{ncpus}{number of cores to use.}
}
\value{
An updated protein dataframe with a Class attribute calculated for
each position.
}
\description{
Takes a protein data frame and
labels each protein position based on known epitope/non-epitope regions
documented in an \strong{epitopes} dataframe.
}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk})
}
