% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_splits.R
\name{make_data_splits}
\alias{make_data_splits}
\title{Split peptide data conditionally on peptide or protein similarity.}
\usage{
make_data_splits(
  df,
  proteins,
  save_folder,
  split_level = "protein",
  split_perc = c(75, 25),
  coverage_threshold = 75,
  identity_threshold = 75,
  ncpus = 1
)
}
\arguments{
\item{df}{data frame of consolidated epitope information, returned by
\code{\link[=extract_peptides]{extract_peptides()}} or \code{\link[=calculate_features]{calculate_features()}}.}

\item{proteins}{data frame of protein data (returned by \code{\link[=get_proteins]{get_proteins()}}).}

\item{save_folder}{path to folder for saving the results.}

\item{split_perc}{numeric vector of desired splitting percentages. See
Details.}

\item{coverage_threshold}{coverage threshold for grouping proteins by
similarity, see Details.}

\item{identity_threshold}{identity threshold for grouping proteins by
similarity, see Details.}

\item{ncpus}{positive integer, number of cores to use}
}
\value{
A list object containing the data splits and additional summary
information.
}
\description{
\strong{NOTE}: this routine requires BLAST+ to be installed in your
local machine. For details on how to set up BLAST+ on your machine, check
\url{https://blast.ncbi.nlm.nih.gov/Blast.cgi?PAGE_TYPE=BlastDocs&DOC_TYPE=Download}.
This function was tested with BLAST versions 2.10.0 and 2.12.0.
}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk})
}
