% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_proteins_dataset.R
\name{make_proteins_dataset}
\alias{make_proteins_dataset}
\title{Builds a partially labelled dataset from protein and epitope sequences}
\usage{
make_proteins_dataset(
  epitopes,
  proteins,
  taxonomy_list,
  prot_IDs,
  orgIDs = NULL,
  removeIDs = NULL,
  hostIDs = NULL,
  min_epit = 8,
  max_epit = 25,
  only_exact = FALSE,
  pos.mismatch.rm = "all",
  set.positive = "mode",
  window_size = 2 * min_epit - 1,
  max.N = 2,
  save_folder = "./",
  ncpus = 1
)
}
\arguments{
\item{epitopes}{data frame of epitope data (returned by \code{\link[=get_LBCE]{get_LBCE()}}).}

\item{proteins}{data frame of protein data (returned by \code{\link[=get_proteins]{get_proteins()}}).}

\item{taxonomy_list}{list containing taxonomy information
(generated by \code{\link[=get_taxonomy]{get_taxonomy()}})}

\item{prot_IDs}{IDs of proteins to be extracted for the data set}

\item{orgIDs}{vector of organism/taxon IDs to retain (see
\code{\link[=filter_epitopes]{filter_epitopes()}}). If \code{NULL} then no organism ID  filtering is
performed.}

\item{removeIDs}{vector of organism IDs to remove (see \code{\link[=filter_epitopes]{filter_epitopes()}}).
Useful for, e.g., using a Class-level \code{orgIDs} and removing some species
or genera. If \code{NULL} then no removal is performed.}

\item{hostIDs}{vector of host IDs to retain  (see \code{\link[=filter_epitopes]{filter_epitopes()}}).
If \code{NULL} then no host filtering is performed.}

\item{min_epit}{positive integer, shortest epitope to be considered}

\item{max_epit}{positive integer, longest epitope to be considered}

\item{only_exact}{logical, should only sequences labelled as "Exact Epitope"
in variable \emph{epit_struc_def} (within \code{epitopes}) be considered?}

\item{pos.mismatch.rm}{should epitopes with position mismatches be removed?
Use "all" (default) for removing any position mismatch or "align"
if the routine should attempt to
search the epitope sequence in the protein sequence.}

\item{set.positive}{how to decide whether an observation should be of the
"Positive" (+1) class? Use "any" to set a sequence as positive if
$n_positive > 0$, "mode" to set it if $n_positive >= n_negative$,
or "all" to set it if $n_negative == 0$. Defaults to "mode".}

\item{window_size}{positive integer, size of window to use.}

\item{max.N}{maximum length of N-peptide frequency features to be calculated.}

\item{save_folder}{path to folder for saving the results.}

\item{ncpus}{number of cores to use for data windowing and feature
calculation.}
}
\value{
Data frame containing the resulting dataset.
}
\description{
This function extracts proteins sequences and labels them based on IEDB
epitope data.
}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk})
}
