% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_window_df.R
\name{make_window_df}
\alias{make_window_df}
\title{Assemble a sliding window representation of epitope data}
\usage{
make_window_df(
  epitopes,
  proteins,
  save_folder = NULL,
  min_epit = 7,
  max_epit = NULL,
  window_size = NULL,
  only_exact = FALSE,
  ncpus = 1
)
}
\arguments{
\item{epitopes}{data frame of epitope data (returned by `get_LBCE()`).}

\item{proteins}{data frame of protein data (returned by `get_proteins()`).}

\item{save_folder}{path to folder for saving the results.}

\item{min_epit}{positive integer, smallest epitope to be considered}

\item{max_epit}{positive integer, largest epitope to be considered}

\item{window_size}{positive integer, size of window to use (see `Details`)}

\item{only_exact}{logical, should only "Exact epitopes" be considered?}

\item{ncpus}{positive integer, number of cores to use}
}
\value{
A data.table object containing the sliding window representation of
the epitope data.
}
\description{
This function is used to turn epitope data into a dataframe where
each row corresponds to a fixed-length window that slides over the epitope
data.
}
\details{
Window size can be set either by the user or automatically. In the latter
case the window size is set as `max(7, (2 * min_epit) - 1)`.

The sliding window runs from the first to the last positions of each entry
in the `epitopes` input (which in turn comes from running `get_LBCE()`).
}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk})
}
