% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_prepare_join_df.R
\name{prepare_join_df}
\alias{prepare_join_df}
\title{Merge and filter epitope and protein data}
\usage{
prepare_join_df(
  epitopes,
  proteins,
  save_folder = NULL,
  min_epit = 8,
  max_epit = 25,
  only_exact = FALSE,
  pos.mismatch.rm = c("all", "align"),
  set.positive = c("any", "mode", "all")
)
}
\arguments{
\item{epitopes}{data frame of epitope data (returned by `get_LBCE()`).}

\item{proteins}{data frame of protein data (returned by running, e.g.,
`get_proteins(unique(epitopes$protein_id))`).}

\item{save_folder}{path to folder for saving the results.}

\item{min_epit}{positive integer, shortest epitope to be considered}

\item{max_epit}{positive integer, longest epitope to be considered}

\item{only_exact}{logical, should only sequences labelled as "Exact Epitope"
in variable `epit_struc_def` be considered?}

\item{pos.mismatch.rm}{should epitopes with position mismatches be removed?
Use `all` (default) for removing any position mismatch (i.e., if the
protein substring between `start_pos` and `end_pos` does not
correspond to `epit_seq`) or `align` if the routine should attempt to
fix `start_pos` and `end_pos` by searching for `epit_seq` in the
protein sequence. Removed entries are listed (if `save_folder` is not
`NULL`) in the `df_errlist` file.}

\item{set.positive}{how to decide whether an observation should be of the
"Positive" class? Use "any" to set a sequence as positive if
`n_positive > 0`, "mode" to set it if `n_positive >= n_negative`,
or "all" to set it if `n_negative == 0`. Defaults to "mode".}
}
\value{
A data.table object containing the merged data frame
}
\description{
Merges previously extracted epitopes and proteins, and performs some
sanity checks and filtering.
}
\details{
Entries in the `epitopes` input are removed if they:
\itemize{
  \item Lack a valid protein ID (i.e., one that has a corresponding entry in
  `proteins`)
  \item Lack a valid string in `epit_seq`
  \item Lack a valid definition in `epit_struc_def`
  \item Have sequences shorter than `min_epit` or longer than `max_epit`.
  \item Have a mismatch between the sequence in `epit_seq` and the
  corresponding sequence between `start_pos` and `end_pos` on the protein
  sequence (see description of parameter `pos.mismatch.rm`).
}
}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk})
}
