% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_protein_for_prediction.R
\name{prepare_protein_for_prediction}
\alias{prepare_protein_for_prediction}
\title{Assemble a sliding window representation of an unlabeled protein}
\usage{
prepare_protein_for_prediction(
  proteins,
  window_size,
  save_folder = NULL,
  ncpus = 1
)
}
\arguments{
\item{proteins}{data frame of protein data (returned by [get_proteins()].}

\item{window_size}{positive integer, size of window to use (see `Details`)}

\item{save_folder}{path to folder for saving the results.}

\item{ncpus}{positive integer, number of cores to use (multi-core
capabilities not yet available for Windows systems.)}
}
\value{
A data frame containing the extracted windows is returned invisibly.
Each row of the resulting data frame will have the protein ID,
windowed sequence, and other variables that may be useful for
mapping each window to its position on the protein sequence.
}
\description{
This function is used to turn protein sequences into windowed
representations, with rows corresponding to a fixed-length window that slides
over the protein. This function is analogous to
`assemble_windowed_dataframe()`, but it works with only a `protein` dataframe
(generated by `get_proteins()`).
}
\details{
If the windowed dataframe is to be used for prediction, it is important to
use the same `window_size` that was used to assemble the training dataset in
`assemble_windowed_dataframe()`, so that equivalent features can be
calculated (notice that sometimes `window_size` is determined automatically
in `assemble_windowed_dataframe()` as `max(3, (2 * min_epit) - 1)`).
}
\examples{
proteins <- data.frame(molecule_id = c("P.001", "A.002"),
              TSeq_sequence = c("TYGACPKYVKQNTLKLATLKLATGMRNVPEKQT",
                "LPKEEKKDDPPKDPKKDDPPKEAQNKLNQPVVADENVDQGPGAPQGPGAPQGPGAPQGPGAPQGPGAPQGPGAPQ"),
              TSeq_taxid = c(NA,NA),
              TSeq_orgname = c(NA,NA),
              stringsAsFactors = FALSE)

x <- prepare_protein_for_prediction(proteins, window_size = 15)

}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk})
}
