% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bepipred2.R
\name{read_bepipred2}
\alias{read_bepipred2}
\title{Read saved output from Bepipred2}
\usage{
read_bepipred2(res.file, threshold = 0.5, ...)
}
\arguments{
\item{res.file}{path to the results file. Must be a CSV file saved from the
Bepipred 2.0 output.}

\item{threshold}{threshold for converting prediction probabilities to
positive class predictions.}

\item{...}{Currently unused.}
}
\description{
This function is used to read the saved CSV output from Bepipred2.
}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk})
}
