% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ibceel.R
\name{read_ibceel}
\alias{read_ibceel}
\title{Read saved output from iBCE-EL}
\usage{
read_ibceel(res.file, ...)
}
\arguments{
\item{res.file}{path to the results file. Must be an HTML file containing the
iBCE-EL output.}

\item{...}{Currently unused.}
}
\description{
This function is used to read the saved HTML output from ABCPred.
}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk})
}
