% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lbtope.R
\name{read_lbtope}
\alias{read_lbtope}
\title{Read saved output from LBtope}
\usage{
read_lbtope(res.file, threshold = 0.5, ...)
}
\arguments{
\item{res.file}{path to the results file. Must be a TXT file saved from the
LBtope output.}

\item{threshold}{threshold for converting prediction probabilities to
positive class predictions.}

\item{...}{Currently unused.}
}
\description{
This function is used to read the saved TXT output from LBtope
}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk})
}
