% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_deprecated.R
\name{retrieve_protein_data}
\alias{retrieve_protein_data}
\title{Retrieve protein sequences and data from GenBank}
\usage{
retrieve_protein_data(uids, save_folder = NULL)
}
\arguments{
\item{uids}{A list of potein IDs provided a character vector.}

\item{save_folder}{path to folder for saving the results.}
}
\description{
This function is deprecated. Please use get_proteins() instead.
}
