% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_predictions.R
\name{smooth_predictions}
\alias{smooth_predictions}
\title{Smoothing of prediction vectors}
\usage{
smooth_predictions(x, window_size, type = "mode")
}
\arguments{
\item{x}{vector of predictions (must be numeric)}

\item{window_size}{size of the sliding window}

\item{type}{type of smoothing function to use. Accepts "mode" (for majority
vote) or "mean".}
}
\value{
vector of smoothed prediction values.
}
\description{
This function runs a sliding window smoother on prediction vectors.
}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk})
}
