% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_windowed_dataframe.R
\name{assemble_windowed_dataframe}
\alias{assemble_windowed_dataframe}
\title{Assemble a sliding window representation of epitope data}
\usage{
assemble_windowed_dataframe()
}
\description{
This function is used to turn epitope and protein data into a dataframe where
each row corresponds to a fixed-length window that slides over the epitope
data (and its surrounding aminoacid residues in the protein).
}
\details{
The initial position of the window is determined as the first position where
the majority of the AAs covered by the window belong to the target region.
Similarly, the last position of the window is determined as the last one
where the majority of the AAs covered by the window still belong to the
target region.

As an example, assume an epitope "QGPGAPQGPGAP" contained within a given
protein, a window size of 9 and a step size of 2. Let "X" represent any AA
not belonging to the target region, and | be a start and end positions covered
by the window. This routine would then generate the following windows (each
in a distinct row of the resulting dataframe):

...XXXXXQGPGAPQGPGAPXXXXX...\cr
. . . | . . . . . . . . . . . . | . . . . . . . . . . . . . . . . . . . XXXXQGPGA\cr
. . . . . . | . . . . . . . . . . . . | . . . . . . . . . . . . . . . . XXQGPGAPQ\cr
. . . . . . . . . | . . . . . . . . . . . . | . . . . . . . . . . . . . QGPGAPQGP\cr
. . . . . . . . . . . . | . . . . . . . . . . . . | . . . . . . . . . . PGAPQGPGA\cr
. . . . . . . . . . . . . . . | . . . . . . . . . . . . | . . . . . . . APQGPGAPX\cr
. . . . . . . . . . . . . . . . . . | . . . . . . . . . . . . | . . . . QGPGAPXXX\cr

Each resulting row of the data frame will have the epitope ID, protein ID,
windowed sequence, and
}
\author{
Felipe Campelo (\email{fcampelo@ufmg.br},
\email{f.campelo@aston.ac.uk})
}
