% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_features.R
\name{calc_features}
\alias{calc_features}
\title{Calculate statistical and physicochemical features for peptides}
\usage{
calc_features(input, max.N = 2, ncores = 1)
}
\arguments{
\item{input}{either a vector of peptides or a data frame with a variable
called "window_seq" containing the peptides.}

\item{max.N}{maximum length of N-peptide frequency features to be calculated.
See \code{\link[=calc_Npeptide_composition]{calc_Npeptide_composition()}} for details.}

\item{ncores}{number of cores to use for calculating some of the features.}
}
\value{
Data frame containing the calculated features. If \strong{input} is
a \emph{data.table} the original input is returned with the features
appended as columns.
}
\description{
This function is used to calculate several distinct families of features for
a vector of peptides.
}
\details{
To get a list of all routines for calculating the
individual types of features use \code{\link[=get_feature_functions]{get_feature_functions()}}. Some of the
features are calculated based on an aminoacid propensity scale, which was
originally obtained from \emph{E.L. Ulrich et al., "BioMagResBank". Nucleic Acids
Research 36, D402-D408 (2008) DOI: 10.1093/nar/gkm957}. The data was
downloaded from \url{http://www.bmrb.wisc.edu/ref_info/aadata.dat}
}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk});
Jodie Ashford (\email{ashfojsm@aston.ac.uk})
}
