% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_features.R
\name{calc_features}
\alias{calc_features}
\title{Calculate statistical and physicochemical features for peptides}
\usage{
calc_features(df, max.N = 2, ncpus = 1, save_folder = NULL)
}
\arguments{
\item{df}{a \emph{data.table} of class \emph{windowed_epit_dt} or \emph{windowed_prot_dt},
generated by \code{\link[=make_window_df]{make_window_df()}}}

\item{max.N}{maximum length of N-peptide frequency features to be calculated.}

\item{ncpus}{number of cores to use.}

\item{save_folder}{path to folder for saving the output.}
}
\value{
Data table containing the calculated features appended as new columns.
}
\description{
This function is used to calculate several distinct families of features for
a vector of peptides.
}
\details{
This function is partially based on functions \code{\link[Peptides:aaComp]{Peptides::aaComp()}}
and \code{\link[Peptides:aaDescriptors]{Peptides::aaDescriptors()}} (check documentation for details).

Some of the features are calculated based on an aminoacid propensity scale,
which was originally obtained from \emph{E.L. Ulrich et al., "BioMagResBank".
Nucleic Acids Research 36, D402-D408 (2008) DOI: 10.1093/nar/gkm957}. The
data was downloaded from
\url{http://www.bmrb.wisc.edu/ref_info/aadata.dat}
on July 1, 2020.
}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk});
Jodie Ashford (\email{ashfojsm@aston.ac.uk})
}
