% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_features.R
\name{calc_features}
\alias{calc_features}
\title{Calculate features for epitope prediction}
\usage{
calc_features(
  splits.list,
  peptide.features = list(),
  protein.features = list(),
  ncpus = 1
)
}
\arguments{
\item{splits.list}{list object returned by \code{\link[=make_data_splits]{make_data_splits()}}.}
}
\description{
This function is used to calculate several distinct families of features for
epitope prediction. Two major types of features can be calculated:
}
\details{
\itemize{
\item \emph{local features}, which are calculated based on the local
neighbourhood of each AA position (column \emph{Info_window} of the windowed
data frame).
\item \emph{global features}, which are calculated using the full sequence
of the protein (listed in column \emph{Info_protein_id}  of the windowed
data frame)
}

All features are calculated based on the implementations available in package
\href{https://cran.r-project.org/package=protr}{\strong{protr}}.
}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk})
}
