% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_molecular_weight.R
\name{calc_molecular_weight}
\alias{calc_molecular_weight}
\title{Calculate molecular weight of peptide sequences}
\usage{
calc_molecular_weight(df, cl)
}
\arguments{
\item{df}{a \emph{data.table} of class \emph{windowed_epit_dt} or \emph{windowed_prot_dt},
generated by \code{\link[=make_window_df]{make_window_df()}}}

\item{cl}{either a cluster object or a positive integer.}
}
\value{
updated \strong{df} data.table containing the molecular weight features
appended as columns.
}
\description{
This function is used to calculate the molecular weights for each peptide
}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk});
Jodie Ashford (\email{ashfojsm@aston.ac.uk})
}
