% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_Npeptide_composition.R
\name{calc_Npeptide_composition}
\alias{calc_Npeptide_composition}
\title{Calculate frequencies of N-peptides}
\usage{
calc_Npeptide_composition(input, N = 2, ncores = 1)
}
\arguments{
\item{input}{either a vector of peptides or a data frame with a variable
called "window_seq" containing the peptides.}

\item{N}{the length of the subsequences to consider.}

\item{ncores}{number of cores to use when calculating the features.}
}
\value{
Data frame containing the calculated AA percentages. If \code{input} is
a\code{data.frame} the original input is returned with the AA percentages
appended as columns.
}
\description{
This function is used to calculate the frequencies of N-peptide subsequences
from a vector of peptides.
}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk});
Jodie Ashford (\email{ashfojsm@aston.ac.uk})
}
