% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_number_of_atoms.R
\name{calc_number_of_atoms}
\alias{calc_number_of_atoms}
\title{Calculate number of each type of atom in peptide sequences}
\usage{
calc_number_of_atoms(input)
}
\arguments{
\item{input}{either a vector of peptides or a data frame with a variable
called "window_seq" containing the peptides.}
}
\value{
Data frame containing the calculated numbers of atoms. If `input` is
a`data.frame` the original input is returned with the numbers of atoms
appended as columns.
}
\description{
This function is used to calculate the number of specific atoms from a vector
of peptides.
}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk});
        Jodie Ashford (\email{ashfojsm@aston.ac.uk})
}
