% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_performance.R
\name{calc_performance}
\alias{calc_performance}
\title{Calculate epitope classification performance indices}
\usage{
calc_performance(truth, pred, prob = NULL, posValue = 1, negValue = -1)
}
\arguments{
\item{truth}{vector of reference values}

\item{pred}{vector of predicted values}

\item{prob}{optional, vector of predicted probabilities}

\item{posValue}{value that indicates the "positive" class.}

\item{negValue}{value that indicates the "negative" class.}
}
\value{
list containing several performance indicators
}
\description{
This function is used to calculate several distinct BINARY classification
performance indicators. \code{NA} values are not considered in the calculations.
}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk})
}
