% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sequence_entropy.R
\name{calc_sequence_entropy}
\alias{calc_sequence_entropy}
\title{Calculate sequence entropy of peptides}
\usage{
calc_sequence_entropy(df, ncpus)
}
\arguments{
\item{df}{a \emph{data.table} of class \emph{windowed_epit_dt} or \emph{windowed_prot_dt},
generated by \code{\link[=make_window_df]{make_window_df()}}}

\item{ncpus}{positive integer, number of cores to use}
}
\value{
updated \strong{df} data.table containing the sequence entropy of each
peptide
}
\description{
This function is used to calculate the sequence entropies,
Entropy = -sum(p_k * log2(p_k))
}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk});
Jodie Ashford (\email{ashfojsm@aston.ac.uk})
}
