% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_peptides.R
\name{extract_peptides}
\alias{extract_peptides}
\title{Extract peptides and windowed representation for epitope prediction.}
\usage{
extract_peptides(
  df,
  min_peptide = 8,
  max_epitope = 30,
  window_size = (2 * min_peptide) - 1,
  save_folder = NULL
)
}
\arguments{
\item{df}{data frame of consolidated protein-epitope data, returned by
\code{\link[=consolidate_data]{consolidate_data()}}.}

\item{min_peptide}{positive integer, shortest peptide to be considered
(applies to both \strong{positive} and \strong{negative} observations)}

\item{max_epitope}{positive integer, longest peptide to be considered (only
applies to \strong{positive} observations.)}

\item{window_size}{positive integer, size of the local neighbourhood to be
considered.}

\item{save_folder}{path to folder for saving the results.}
}
\value{
List containing two data frames:

\itemize{
\item A data frame containing the labeled positions of \code{df} (filtered by
\emph{min_epitope} and \emph{max_epitope}) with a new column expressing the local
neighbourhood of each position (to be used for feature calculation)
\item A data frame containing the individual peptides and classes
}
}
\description{
Extract relevant peptides from the consolidated protein-epitope data
(generated using \code{\link[=consolidate_data]{consolidate_data()}}) and build column with neightbourhood
of each labeled aminoacid residue.
}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk})
}
