% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_epitopes.R
\name{filter_epitopes}
\alias{filter_epitopes}
\title{Filter epitopes by taxonomy and host}
\usage{
filter_epitopes(
  df,
  orgIDs = NULL,
  hostIDs = NULL,
  removeIDs = NULL,
  tax_load_file = NULL,
  tax_save_folder = NULL
)
}
\arguments{
\item{df}{`data.table` object of class `LBCE_dt` (generated by `get_LBCE()`),
`joined_epit_dt` (generated by `prepare_join_df()`) or
`windowed_epit_dt` (generated by `make_window_df()`)}

\item{orgIDs}{vector of organism taxon IDs that we want to retain
(using `sourceOrg_id`).}

\item{hostIDs}{vector of host taxon IDs that we want to retain
(using `host_id`).}

\item{removeIDs}{vector of organism taxon IDs that we want to remove
(using `sourceOrg_id`).}

\item{tax_load_file}{optional taxonomy file (RDS file generated either by this
function or by [get_taxonomy()]).}

\item{tax_save_folder}{optional folder saving the taxonomy. Ignored
if `tax_load_file` is not `NULL`.}
}
\value{
Epitope dataframe filtered by the criteria in `orgIDs`, `hostIDs` and
        `removeIDs`.
}
\description{
Filters an epitope dataframe (returned by [get_LBCE()] or [get_LTCE()]) by
a specific taxon and/or by an specific host.
}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk})
}
