% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_epitopes.R
\name{filter_epitopes}
\alias{filter_epitopes}
\title{Filter epitopes by taxonomy and host}
\usage{
filter_epitopes(
  df,
  orgIDs = NULL,
  hostIDs = NULL,
  removeIDs = NULL,
  orgID_column = "sourceOrg_id",
  hostID_column = "host_id",
  tax_list = NULL
)
}
\arguments{
\item{df}{a data frame containing epitope data}

\item{orgIDs}{vector of organism taxon IDs that we want to retain
(using \code{df$sourceOrg_id}).}

\item{hostIDs}{vector of host taxon IDs that we want to retain
(using \code{df$host_id}).}

\item{removeIDs}{vector of organism taxon IDs that we want to remove
(using \code{df$sourceOrg_id}).}

\item{orgID_column}{name of the column in \code{df} containing the source organism IDs}

\item{hostID_column}{name of the column in \code{df} containing the host IDs}

\item{tax_list}{taxonomy list (generated by \code{\link[=get_taxonomy]{get_taxonomy()}})}
}
\value{
data frame with epitope data, filtered by the criteria in \strong{orgIDs},
\strong{hostIDs} and \strong{removeIDs}.
}
\description{
Filters an epitope dataframe by source organism and/or host IDs.
}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk})
}
