% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_epitopes.R
\name{filter_epitopes}
\alias{filter_epitopes}
\title{Filter epitopes by taxonomy and host}
\usage{
filter_epitopes(
  df,
  orgIDs = NULL,
  hostIDs = NULL,
  removeIDs = NULL,
  tax_load_file = NULL,
  tax_list = NULL,
  tax_save_folder = NULL
)
}
\arguments{
\item{df}{\emph{data.table} object of class \emph{LBCE_dt} (generated by \code{\link[=get_LBCE]{get_LBCE()}}),
\emph{joined_epit_dt} (generated by \code{\link[=prepare_join_df]{prepare_join_df()}}) or
\emph{windowed_epit_dt} (generated by \code{\link[=make_window_df]{make_window_df()}})}

\item{orgIDs}{vector of organism taxon IDs that we want to retain
(using \emph{sourceOrg_id}).}

\item{hostIDs}{vector of host taxon IDs that we want to retain
(using \emph{host_id}).}

\item{removeIDs}{vector of organism taxon IDs that we want to remove
(using \emph{sourceOrg_id}).}

\item{tax_load_file}{optional taxonomy file (RDS file generated either by this
function or by \code{\link[=get_taxonomy]{get_taxonomy()}}).}

\item{tax_list}{optional taxonomy list (to be passed instead of tax_load_file)}

\item{tax_save_folder}{optional folder saving the taxonomy. Ignored
if \strong{tax_load_file} is not \emph{NULL}.}
}
\value{
Epitope dataframe filtered by the criteria in \strong{orgIDs}, \strong{hostIDs} and
\strong{removeIDs}.
}
\description{
Filters an epitope dataframe by source organism and/or host IDs.
}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk})
}
