% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_epitopes.R
\name{filter_epitopes}
\alias{filter_epitopes}
\title{Filter epitopes by taxonomy and host}
\usage{
filter_epitopes(
  epitopes,
  taxID = NULL,
  hostID = NULL,
  removeID = NULL,
  tax_load_file = NULL,
  tax_save_folder = NULL
)
}
\arguments{
\item{epitopes}{data frame of epitope data (returned by [get_LBCE()] or
[get_LTCE()].}

\item{taxID}{taxon IDs that we want to filter the epitopes by.}

\item{hostID}{host IDs that we want to filter the epitopes by.}

\item{removeID}{organism IDs to be removed from the data frame.}

\item{tax_load_file}{optional taxonomy file (RDS file generated either by this
function or by [get_taxonomy()]).}

\item{tax_save_folder}{optional folder saving the taxonomy. Ignored
if `tax_load_file` is not `NULL`.}
}
\value{
Epitope dataframe filtered by the criteria in `taxID`, `hostID` and
        `removeID`.
}
\description{
Filters an epitope dataframe (returned by [get_LBCE()] or [get_LTCE()]) by
a specific taxon and/or by an specific host.
}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk})
}
