% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Fit Random Forest model to epitope data}
\usage{
fit_model(
  data.train,
  data.test = NULL,
  rnd.seed = NULL,
  ncpus = 1,
  threshold = 0.5,
  ...
)
}
\arguments{
\item{data.train}{data frame containing the training data (one or more
numerical predictors and one \strong{Class} attribute).}

\item{data.test}{data frame containing the test data, in the same format as
\code{data.train}.}

\item{rnd.seed}{seed for random number generator}

\item{ncpus}{number of cores to use.}

\item{threshold}{probability threshold for attributing a prediction as
\emph{positive}.}

\item{...}{options to be passed down to the Random Forest implementation (see
\code{\link[ranger:ranger]{ranger::ranger()}} for details)}
}
\value{
List containing the fitted model, predictions for  and several performance indicators.
}
\description{
This function fits a Random Forest model to epitope data (after feature
calculation using \code{\link[=calc_features]{calc_features()}}). Feature column names should start with
\emph{feat_}, and the class attribute should be called \emph{Class}. All other columns
are ignored when fitting the model.
}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk})
}
