% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_feature_functions.R
\name{get_feature_functions}
\alias{get_feature_functions}
\title{Print available feature calculation functions}
\usage{
get_feature_functions()
}
\value{
vector of function names.
}
\description{
Prints the feature calculation functions available in the \strong{epitopes}
package
}
\details{
This routine prints the names of the feature calculation functions available
in the \strong{epitopes} package, which are used within \strong{calc_features()}.
}
\examples{
get_feature_functions()

}
