% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_linear_bcell_epitopes.R
\name{get_linear_bcell_epitopes}
\alias{get_linear_bcell_epitopes}
\title{Extract linear B Cell epitopes from XML files retrieved from IEDB.}
\usage{
get_linear_bcell_epitopes(data_folder = "./", save_file = NULL)
}
\arguments{
\item{data_folder}{path (either relative or absolute) to the directory
containing the XML files}

\item{save_file}{file name for saving the resulting data frame.}
}
\value{
A data frame containing the extracted epitopes is returned invisibly.
}
\description{
This function is used to extract information for **linear B-cell epitopes**
from the XML files exported using the functionality provided by
[IEDB](https://www.iedb.org/).
It assumes that the user has downloaded the _Complete Database Export_ from
the _XML Database Export_ field in IEDB's
[Database Export](https://www.iedb.org/database_export_v3.php) and extracted
it in a given folder, which is passed as an argument to the function.
}
\author{
Felipe Campelo (\email{fcampelo@ufmg.br},
\email{f.campelo@aston.ac.uk})
}
