% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_proteins.R
\name{get_proteins}
\alias{get_proteins}
\title{Retrieve protein sequences and data from GenBank and Uniprot}
\usage{
get_proteins(uids, save_folder = NULL)
}
\arguments{
\item{uids}{A character vector with protein IDs, or a \emph{data.table} object
of class \emph{LBCE_dt}, generated by \code{\link[=get_LBCE]{get_LBCE()}}.}

\item{save_folder}{path to folder for saving the results.}
}
\value{
A list object containing a data frame with the extracted proteins
plus a vector of IDs that were not successfully retrieved.
}
\description{
This function is used to retrieve data from
Genbank's protein database for
given protein IDs. If an ID is not available from Genbank the function will
try to retrieve it from Uniprot.
}
\details{
Queries are processed one by one (rather than in batch) to enable treatment
of individual inconsistencies (e.g., wrong UIDs, queries that return a
different identifier, etc.). This makes this routine substantially slower,
but considerably more robust to errors.
}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk})
}
