% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taxonomy.R
\name{get_taxonomy}
\alias{get_taxonomy}
\title{Retrieve taxonomic classification tables from NCBI.}
\usage{
get_taxonomy(UIDs)
}
\arguments{
\item{UIDs}{vector of organism IDs to retrieve taxonomical information}
}
\value{
A list containing the information for each element of UIDs.
}
\description{
This function retrieves the taxonomic information of a vector of organism
IDs, from the NCBI Taxonomy data base.
}
\examples{
UIDs <- c("6282", # O. volvulus,
          "9606") # H. sapiens
tax <- get_taxonomy(UIDs)

}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk})
}
