% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_splits.R
\name{make_data_splits}
\alias{make_data_splits}
\title{Split peptide data conditionally on peptide or protein similarity.}
\usage{
make_data_splits(
  df,
  peptides = NULL,
  proteins = NULL,
  split_level = "protein",
  split_perc = c(0.75, 0.25),
  similarity_threshold = 0.7,
  save_folder = NULL,
  ncpus = 1,
  substitution_matrix = "BLOSUM62",
  SAopts = list()
)
}
\arguments{
\item{df}{data frame of consolidated epitope information, returned by
\code{\link[=extract_peptides]{extract_peptides()}} or \code{\link[=calculate_features]{calculate_features()}}.}

\item{proteins}{data frame of protein data (returned by \code{\link[=get_proteins]{get_proteins()}}).}

\item{split_perc}{numeric vector of desired splitting percentages. See
Details.}

\item{save_folder}{path to folder for saving the results.}

\item{ncpus}{positive integer, number of cores to use}

\item{coverage_threshold}{coverage threshold for grouping proteins by
similarity, see Details.}

\item{identity_threshold}{identity threshold for grouping proteins by
similarity, see Details.}
}
\value{
A list object containing the data splits and additional summary
information.
}
\description{
Split peptide data conditionally on peptide or protein similarity.
}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk})
}
