% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_join_df.R
\name{prepare_join_df}
\alias{prepare_join_df}
\title{Merge and filter epitope and protein data}
\usage{
prepare_join_df(
  epitopes,
  proteins,
  save_folder = NULL,
  min_epit = 8,
  max_epit = 25,
  only_exact = FALSE,
  pos.mismatch.rm = c("all", "align"),
  set.positive = c("any", "mode", "all")
)
}
\arguments{
\item{epitopes}{data frame of epitope data (returned by \code{\link[=get_LBCE]{get_LBCE()}}).}

\item{proteins}{data frame of protein data (returned by \code{\link[=get_proteins]{get_proteins()}}).}

\item{save_folder}{path to folder for saving the results.}

\item{min_epit}{positive integer, shortest epitope to be considered}

\item{max_epit}{positive integer, longest epitope to be considered}

\item{only_exact}{logical, should only sequences labelled as "Exact Epitope"
in variable \emph{epit_struc_def} (within \code{epitopes}) be considered?}

\item{pos.mismatch.rm}{should epitopes with position mismatches be removed?
Use "all" (default) for removing any position mismatch or "align"
if the routine should attempt to
search the epitope sequence in the protein sequence.}

\item{set.positive}{how to decide whether an observation should be of the
"Positive" (+1) class? Use "any" to set a sequence as positive if
$n_positive > 0$, "mode" to set it if $n_positive >= n_negative$,
or "all" to set it if $n_negative == 0$. Defaults to "mode".}
}
\value{
A data.table object containing the merged data frame
}
\description{
Merges previously extracted epitopes and proteins, and performs some
sanity checks and filtering.
}
\details{
Entries in the \code{epitopes} input are removed if they:
\itemize{
\item Lack a valid protein ID (i.e., one that has a corresponding entry in
\code{proteins})
\item Lack a valid string in \emph{epit_seq}
\item Lack a valid definition in \emph{epit_struc_def}
\item Have sequences shorter than \code{min_epit} or longer than \code{max_epit}.
\item Have a mismatch between the sequence in \emph{epit_seq} and the
corresponding sequence between \emph{start_pos} and \emph{end_pos} on the protein
sequence (see description of parameter \code{pos.mismatch.rm}).
}
}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk})
}
