% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ibceel.R
\name{read_ibceel}
\alias{read_ibceel}
\title{Read saved output from iBCE-EL}
\usage{
read_ibceel(res.file, split_string = "pp", ...)
}
\arguments{
\item{res.file}{path to the results file. Must be an HTML file containing the
iBCE-EL output.}

\item{split_string}{short string used to separate protein ID from position in
the iBCE-EL output. This assumes that the iBCE-EL results were generated
from a FASTA file in which each peptide had the ID string set as
\emph{proteinID<split_string>position}.}

\item{...}{Currently unused.}
}
\description{
This function is used to read the saved HTML output from ABCPred.
}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk})
}
