% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_protein_data.R
\name{retrieve_protein_data}
\alias{retrieve_protein_data}
\title{Retrieve protein sequences and data from GenBank}
\usage{
retrieve_protein_data(uids, save_file)
}
\arguments{
\item{uids}{A list of potein IDs provided a character vector.}

\item{save_file}{file name for saving the resulting data frame.}
}
\value{
A data frame containing the extracted proteins is returned invisibly.
}
\description{
This function is used to retrieve data from
[GenBank](https://www.ncbi.nlm.nih.gov/genbank/) for given protein IDs.
}
\author{
Felipe Campelo (\email{fcampelo@ufmg.br},
\email{f.campelo@aston.ac.uk})
}
