% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_predictions.R
\name{smooth_predictions}
\alias{smooth_predictions}
\title{Smoothing of prediction vectors}
\usage{
smooth_predictions(x, window_size, type = "mode", minsize = 8)
}
\arguments{
\item{x}{vector of predictions (must be numeric)}

\item{window_size}{size of the sliding window. Only used if \code{type} is
either \code{"mean"} or \code{"mode"}.}

\item{type}{type of smoothing function to use. Accepts "mode" (simple
majority vote), "mean" or "minsize".}

\item{minsize}{smallest sequence length for positive predictions.
Stretches of positive-labelled positions shorter than \code{minsize} are
flipped to negative (only works if \verb{type == "minsize}).}
}
\value{
vector of smoothed prediction values.
}
\description{
This function runs a sliding window smoother on prediction vectors.
}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk})
}
